AI project5/requirements.txt
torch>=1.10
torchvision
ftfy
regex
tqdm
scikit-learn
Pillow
git+https://github.com/openai/CLIP.git